Add validation of maximum progress

+++ src/QueReal.BLL/Interfaces/IQuestService.cs
@@ -14,7 +14,7 @@ namespace QueReal.BLL.Interfaces
 
 		Task DeleteAsync(Guid questId);
 
-		Task SetProgressAsync(Guid questItemId, short progress);
+		Task SetProgressAsync(Guid questItemId, byte progress);
 
 		Task ApproveCompletionAsync(Guid questId);
